Refactor: Move code into separate modules
